Min Jumps
Given an array of integers where each element represents the maximum number of steps that can be
made forward from that element, write a function to return the minimum number of jumps to reach
the end of the array, starting from the first element.
If an element is 0, then we cannot move through that element.