String to integer
Implement atoi which converts a string to an integer.
The function first discards as many whitespace characters as necessary
until the first non-whitespace character is found.
Then, starting from this character, takes an optional initial plus or
minus sign followed by as many numerical digits as possible, and interprets them as a numerical value.
If no valid conversion could be performed, a zero value is returned.