dolls
We have a bin of doll parts in a factory. Each part goes to a doll with a specific, unique name.
Each part will be described by a string, with the name of the doll and the part name separated by an underscore, like 'Mary_arm'.
All dolls are made of the same types of parts, and we have a list of all of the parts that form a complete doll.
Given a list of available parts, return a list of doll names for which will we be able to create at least one complete doll.